requirements_deploy.txt
streamlit==1.28.1
google-cloud-aiplatform==1.38.1
langchain-google-vertexai==0.0.6
google-auth==2.23.4 